typemap_ors
Trait CloneAny

pub trait CloneAny:
    Any
    + Send
    + Sync { }

A marker trait meant for use as the A parameter in TypeMap.

This can be used to construct TypeMaps containing only types which implement Clone like so: TypeMap::<CloneAny>::custom(), which produces a TypeMap<CloneAny>. Combine CloneAny with Send or Sync to add additional bounds.

There is also an exported alias for this type of TypeMap, CloneAny.